Unflattened Device Tree/Generic 8250
I put of making any commits for awhile and so two pretty big changes are
being merged with a bunch of smaller ones into one very large commit.

First, the interface to the flat device tree was getting a bit
frustrating, especially when we need to handle phandles to different
devices like interrupt controllers and clocks. So to avoid a large
amount of complicated code being replicated around the kernel, we're now
just "unflattening" the device tree once, and adding an abstraction
layer on top of it. (Plus the same abstraction layer could be placed
ontop of ACPI as is done in Linux to make drivers a bit more flexible).

The other big change is that we need a UART driver for the rockpro,
which is based on the 8250/16550 line of devices, and so I wrote a rough
"generic 8250" framework which any UART device derived from the 8250 can
base themselves on. This should avoid a ton of code replication. But
I've chosen to leave the x86 serial driver alone instead of trying to
put it ontop of this framework too.

(P.S. I've only tested ARM for this commit and I'm 100% sure RISC-V
won't work with it, but 50/50 on whether or not x86 could handle the
changes I made)