Compiling ARM64 with Clang 15 and WLLVM
Only tested with QEMU virt so far, but it's working pretty consistently
now. Hangs on boot every once in awhile but if it makes it to the shell
it functions well. It mostly came down to a few linker script issues and
inline assembly which Clang didn't like but GCC was okay with. Added
-fomit-frame-pointer to Makefile.arm64 to avoid stack corruption when
changing stacks in "init".